src/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
chardet==5.2.0
colorama==0.4.6
config==0.5.1
distlib==0.3.8
Django==4.2.14
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
filelock==3.15.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysqlclient==2.1.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pyproject-api==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.5.0
tomli==2.0.1
tox==4.16.0
typing-extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
